reformat: move config struct to top module
