+++ src/lib.rs
@@ -6,6 +6,7 @@ pub use stream::*;
 #[macro_use]
 mod utils;
 mod builder;
+mod config;
 mod errors;
 pub mod packet;
 mod stream;
